ci: move every Unix-only test behind a file build tag
Six tests kept a Unix fact inside a shared file behind an inline
`runtime.GOOS` check. A new test in the same file has no signal that it
needs the same check. Each one moves to a `_unix_test.go` file, so the
convention is uniform: a shared file runs on every platform, and a
`_unix_test.go` file runs on Unix. A Unix-only test that lands in a shared
file fails the Windows job, which is the signal.
