+++ internal/doctor/checks_test.go
@@ -180,24 +180,6 @@ func TestPrependPath(t *testing.T) {
 	assert.Equal(t, "TERM=xterm", result[2])
 }
 
-func TestCheckShimScripts(t *testing.T) {
-	if runtime.GOOS == "windows" {
-		// CheckShimScripts reads the executable bit, which Windows does not
-		// have. The Windows form of the shim-directory check owns this.
-		t.Skip("executable bit is Unix only")
-	}
-	tmpDir := t.TempDir()
-	shimDir := filepath.Join(tmpDir, ".pmg", "bin")
-	require.NoError(t, os.MkdirAll(shimDir, 0o755))
-
-	shimPath := filepath.Join(shimDir, "npm")
-	require.NoError(t, os.WriteFile(shimPath, []byte("#!/bin/sh\nexec pmg npm \"$@\""), 0o755))
-
-	found, missing := CheckShimScripts(shimDir, []string{"npm", "pip"})
-	assert.Equal(t, []string{"npm"}, found)
-	assert.Equal(t, []string{"pip"}, missing)
-}
-
 func exeSuffix() string {
 	if runtime.GOOS == "windows" {
 		return ".exe"
