+++ internal/fsutil/fsutil_test.go
@@ -3,7 +3,6 @@ package fsutil
 import (
 	"os"
 	"path/filepath"
-	"runtime"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -39,21 +38,4 @@ func TestMkdirAllRootOwnedRejectsFileCollision(t *testing.T) {
 
 	assert.Error(t, MkdirAllRootOwned(blocker, 0o755))
 	assert.Error(t, MkdirAllRootOwned(filepath.Join(blocker, "sub"), 0o755))
-}
-
-func TestMkdirAllRootOwnedLeavesExistingDirsAlone(t *testing.T) {
-	root := t.TempDir()
-	existing := filepath.Join(root, "existing")
-	require.NoError(t, os.Mkdir(existing, 0o700))
-
-	require.NoError(t, MkdirAllRootOwned(filepath.Join(existing, "created"), 0o755))
-
-	if runtime.GOOS == "windows" {
-		// Windows has no Unix permission bits to preserve.
-		return
-	}
-	info, err := os.Stat(existing)
-	require.NoError(t, err)
-	assert.Equal(t, os.FileMode(0o700), info.Mode().Perm(),
-		"pre-existing directory permissions must not be changed")
 }
