+++ internal/proxyserver/state_test.go
@@ -3,7 +3,6 @@ package proxyserver
 import (
 	"os"
 	"path/filepath"
-	"runtime"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -47,17 +46,6 @@ func TestRemoveState(t *testing.T) {
 
 	_, err := os.Stat(path)
 	assert.True(t, os.IsNotExist(err))
-}
-
-func TestIsRunningCurrentProcess(t *testing.T) {
-	if runtime.GOOS == "windows" {
-		// IsRunning probes with Signal(0), which Windows does not support.
-		// The daemon is not supported on Windows, and its liveness check is
-		// its own issue.
-		t.Skip("Signal(0) liveness probe is Unix only")
-	}
-	s := State{PID: os.Getpid(), Addr: "127.0.0.1:1"}
-	assert.True(t, s.IsRunning())
 }
 
 func TestIsRunningDeadPID(t *testing.T) {
