+++ internal/runner/childexit_test.go
@@ -3,7 +3,6 @@ package runner
 import (
 	"errors"
 	"os/exec"
-	"runtime"
 	"testing"
 
 	"github.com/safedep/dry/usefulerror"
@@ -14,28 +13,6 @@ import (
 )
 
 func TestExtractExit(t *testing.T) {
-	t.Run("direct non-zero exit resolves the real code", func(t *testing.T) {
-		requireSh(t)
-		err := exec.Command("sh", "-c", "exit 2").Run()
-		require.Error(t, err)
-
-		code, signaled, resolved := extractExit(err)
-		assert.Equal(t, 2, code)
-		assert.False(t, signaled)
-		assert.True(t, resolved)
-	})
-
-	t.Run("direct signal termination resolves to 128+signum", func(t *testing.T) {
-		requireSh(t)
-		err := exec.Command("sh", "-c", "kill -INT $$").Run()
-		require.Error(t, err)
-
-		code, signaled, resolved := extractExit(err)
-		assert.Equal(t, 130, code) // 128 + SIGINT(2)
-		assert.True(t, signaled)
-		assert.True(t, resolved)
-	})
-
 	t.Run("pty exit error reads its fields directly", func(t *testing.T) {
 		code, signaled, resolved := extractExit(&pty.ExitError{Code: 1})
 		assert.Equal(t, 1, code)
@@ -112,12 +89,4 @@ func TestClassifyTransparentChildExit(t *testing.T) {
 	require.True(t, errors.As(err, &ce))
 	assert.Equal(t, 1, ce.ExitCode())
 	assert.Equal(t, "npm", ce.PMName)
-}
-
-// requireSh skips a case that drives sh and Unix signals.
-func requireSh(t *testing.T) {
-	t.Helper()
-	if runtime.GOOS == "windows" {
-		t.Skip("needs sh and Unix exit semantics")
-	}
 }
